refactor(scraping): Extract policy leaves
