+++ linkedin_mcp_server/scraping/link_metadata.py
@@ -114,7 +114,7 @@ _REFERENCE_CAPS = {
     "inbox": 30,
     "conversation": 12,
     # Headroom for get_feed's num_posts ceiling (Field(ge=1, le=50)).
-    # Kept in sync with the literal cap=50 in extractor._build_feed_references
+    # Kept in sync with the literal cap=50 in feed_payload.build_feed_references
     # where SDUI-derived /posts/<slug> permalinks are appended.
     "feed": 50,
 }
