+++ tests/test_scraping.py
@@ -27,6 +27,7 @@ from linkedin_mcp_server.scraping.connection import (
     detect_connection_state,
 )
 from linkedin_mcp_server.scraping import extractor as extractor_module
+from linkedin_mcp_server.scraping.contracts import RATE_LIMITED_SECTION_TEXT
 from linkedin_mcp_server.scraping.extractor import (
     ExtractedSection,
     LinkedInExtractor,
@@ -35,11 +36,6 @@ from linkedin_mcp_server.scraping.extractor import (
     _MESSAGE_CONFIRMATION_DISPOSE_JS,
     _MESSAGE_CONFIRMATION_PREPARE_JS,
     _MESSAGE_CONFIRMATION_READY_JS,
-    _RATE_LIMITED_MSG,
-    _build_feed_references,
-    _truncate_linkedin_noise,
-    strip_conversation_chrome,
-    strip_linkedin_noise,
 )
 from linkedin_mcp_server.scraping.link_metadata import Reference
 
@@ -382,7 +378,7 @@ class TestExtractPage:
                 section_name="experience",
             )
 
-        assert result.text == _RATE_LIMITED_MSG
+        assert result.text == RATE_LIMITED_SECTION_TEXT
         # goto called twice (initial + retry)
         assert mock_page.goto.await_count == 2
 
@@ -2975,7 +2971,7 @@ class TestConnectWithPerson:
                 "extract_page",
                 new_callable=AsyncMock,
                 side_effect=[
-                    extracted(_RATE_LIMITED_MSG),
+                    extracted(RATE_LIMITED_SECTION_TEXT),
                     extracted("Post text"),
                 ],
             ) as mock_extract,
@@ -3012,7 +3008,7 @@ class TestConnectWithPerson:
                 extractor,
                 "extract_page",
                 new_callable=AsyncMock,
-                return_value=extracted(_RATE_LIMITED_MSG),
+                return_value=extracted(RATE_LIMITED_SECTION_TEXT),
             ),
             patch.object(
                 extractor,
@@ -3040,7 +3036,7 @@ class TestConnectWithPerson:
                 new_callable=AsyncMock,
                 side_effect=[
                     extracted("Profile text"),
-                    extracted(_RATE_LIMITED_MSG),
+                    extracted(RATE_LIMITED_SECTION_TEXT),
                 ],
             ),
             patch.object(
@@ -3140,7 +3136,7 @@ class TestScrapeCompany:
                 "extract_page",
                 new_callable=AsyncMock,
                 side_effect=[
-                    extracted(_RATE_LIMITED_MSG),
+                    extracted(RATE_LIMITED_SECTION_TEXT),
                     extracted("Posts text"),
                 ],
             ) as mock_extract,
@@ -3243,7 +3239,7 @@ class TestScrapeJob:
             extractor,
             "extract_page",
             new_callable=AsyncMock,
-            return_value=extracted(_RATE_LIMITED_MSG),
+            return_value=extracted(RATE_LIMITED_SECTION_TEXT),
         ):
             result = await extractor.scrape_job("12345")
 
@@ -3876,7 +3872,7 @@ class TestSearchJobs:
         pages = iter(
             [
                 extracted("python jobs"),
-                extracted(_RATE_LIMITED_MSG),
+                extracted(RATE_LIMITED_SECTION_TEXT),
             ]
         )
         urls = iter(
@@ -3916,7 +3912,7 @@ class TestSearchJobs:
         assert result["job_ids"] == ["901"]
         message = result["section_errors"]["search_results"]["error_message"]
         assert "location" in message
-        assert _RATE_LIMITED_MSG in message
+        assert RATE_LIMITED_SECTION_TEXT in message
 
     async def test_a_search_answered_for_something_else_stops_the_loop(self, mock_page):
         """The route can be right and the offset right while the query is gone.
@@ -5020,7 +5016,7 @@ class TestSearchJobs:
                 extractor,
                 "_extract_search_page",
                 new_callable=AsyncMock,
-                return_value=extracted(_RATE_LIMITED_MSG),
+                return_value=extracted(RATE_LIMITED_SECTION_TEXT),
             ),
             patch.object(
                 extractor,
@@ -5055,7 +5051,7 @@ class TestSearchJobs:
                 "_extract_search_page",
                 side_effect=self._navigating(
                     mock_page,
-                    [extracted(_RATE_LIMITED_MSG)],
+                    [extracted(RATE_LIMITED_SECTION_TEXT)],
                     lands_on="https://www.linkedin.com/feed/",
                 ),
             ),
@@ -5863,7 +5859,7 @@ class TestGetSavedJobs:
         were no more pages" — which look identical otherwise.
         """
         extractor = LinkedInExtractor(mock_page)
-        pages = iter([extracted("first page"), extracted(_RATE_LIMITED_MSG)])
+        pages = iter([extracted("first page"), extracted(RATE_LIMITED_SECTION_TEXT)])
         with (
             patch.object(
                 extractor,
@@ -6040,7 +6036,7 @@ class TestSingleSectionRateLimits:
             extractor,
             "extract_page",
             new_callable=AsyncMock,
-            return_value=extracted(_RATE_LIMITED_MSG),
+            return_value=extracted(RATE_LIMITED_SECTION_TEXT),
         ):
             result = await getattr(extractor, method)(*args)
 
@@ -6274,7 +6270,7 @@ class TestSearchPosts:
             extractor,
             "extract_page",
             new_callable=AsyncMock,
-            return_value=extracted(_RATE_LIMITED_MSG),
+            return_value=extracted(RATE_LIMITED_SECTION_TEXT),
         ):
             result = await extractor.search_posts("python")
 
@@ -6298,208 +6294,6 @@ class TestSearchPosts:
             "error_type": "navigation_error",
             "error_message": "timeout",
         }
-
-
-class TestStripLinkedInNoise:
-    def test_strips_footer(self):
-        text = "Bill Gates\nChair, Gates Foundation\n\nAbout\nAccessibility\nTalent Solutions\nCareers"
-        assert strip_linkedin_noise(text) == "Bill Gates\nChair, Gates Foundation"
-
-    def test_strips_footer_with_talent_solutions_variant(self):
-        text = "Profile content here\n\nAbout\nTalent Solutions\nMore footer"
-        assert strip_linkedin_noise(text) == "Profile content here"
-
-    def test_strips_sidebar_recommendations(self):
-        text = "Experience\nCo-chair\nGates Foundation\n\nMore profiles for you\nSundar Pichai\nCEO at Google"
-        assert strip_linkedin_noise(text) == "Experience\nCo-chair\nGates Foundation"
-
-    def test_strips_premium_upsell(self):
-        text = "Education\nHarvard University\n\nExplore premium profiles\nRandom Person\nSoftware Engineer"
-        assert strip_linkedin_noise(text) == "Education\nHarvard University"
-
-    def test_picks_earliest_marker(self):
-        text = "Content\n\nExplore premium profiles\nStuff\n\nMore profiles for you\nMore stuff\n\nAbout\nAccessibility"
-        assert strip_linkedin_noise(text) == "Content"
-
-    def test_no_noise_returns_unchanged(self):
-        text = "Clean content with no LinkedIn chrome"
-        assert strip_linkedin_noise(text) == "Clean content with no LinkedIn chrome"
-
-    def test_empty_string(self):
-        assert strip_linkedin_noise("") == ""
-
-    def test_truncate_noise_preserves_media_controls_for_rate_limit_detection(self):
-        text = "Play\nLoaded: 100.00%\nRemaining time 0:07\nShow captions"
-        assert _truncate_linkedin_noise(text) == text
-        assert strip_linkedin_noise(text) == ""
-
-    def test_about_in_profile_content_not_stripped(self):
-        """'About' followed by actual content (not 'Accessibility') should be preserved."""
-        text = "About\nChair of the Gates Foundation.\n\nFeatured\nPost"
-        assert (
-            strip_linkedin_noise(text)
-            == "About\nChair of the Gates Foundation.\n\nFeatured\nPost"
-        )
-
-    def test_real_footer_with_languages(self):
-        text = (
-            "Company info\n\n"
-            "About\nAccessibility\nTalent Solutions\nCareers\n"
-            "Select language\nEnglish (English)\nDeutsch (German)"
-        )
-        assert strip_linkedin_noise(text) == "Company info"
-
-    def test_preserves_real_careers_content(self):
-        text = "Careers\nWe're hiring globally.\nOpen roles in engineering and design."
-        assert strip_linkedin_noise(text) == text
-
-    def test_preserves_real_questions_content(self):
-        text = "Questions?\nReach out to our recruiting team for details."
-        assert strip_linkedin_noise(text) == text
-
-    def test_strips_media_controls_lines(self):
-        text = (
-            "Feed post number 1\n"
-            "Play\n"
-            "Loaded: 100.00%\n"
-            "Remaining time 0:07\n"
-            "Playback speed\n"
-            "Actual post content\n"
-            "Show captions\n"
-            "Close modal window"
-        )
-        assert strip_linkedin_noise(text) == "Feed post number 1\nActual post content"
-
-
-class TestStripConversationChrome:
-    THREAD = (
-        "MAY 25\n"
-        "Grace Hopper sent the following message at 5:27 PM\n"
-        "Grace Hopper  5:27 PM\n"
-        "\n"
-        "Hello!"
-    )
-    PAGE = (
-        "Messaging\n"
-        "Search messages\n"
-        "Compose a new message\n"
-        "Inbox\n"
-        "Attention screen reader users, messaging items continuously update.\n"
-        "Ada Lovelace\n"
-        "Jun 8\n"
-        "Ada: Preview belonging to a different conversation\n"
-        ". Press return to go to conversation details\n"
-        "Open the options list in your conversation with Ada Lovelace and Grace Hopper\n"
-        "Status is reachable\n"
-        "Load more conversations\n"
-        "Grace Hopper\n"
-        "Status is online\n"
-        "Open the options list in your conversation with Grace Hopper and Ada Lovelace\n"
-        + THREAD
-        + "\n"
-        "Maximize compose field\n"
-        "Attach an image to your conversation with Grace Hopper\n"
-        "Open GIF Keyboard\n"
-        "Send\n"
-        "Open send options"
-    )
-
-    def test_strips_sidebar_and_composer(self):
-        assert strip_conversation_chrome(self.PAGE) == self.THREAD
-
-    def test_other_conversation_previews_removed(self):
-        assert "different conversation" not in strip_conversation_chrome(self.PAGE)
-        assert "Ada Lovelace" not in strip_conversation_chrome(self.PAGE)
-
-    def test_missing_composer_strips_only_leading_chrome(self):
-        text = (
-            "Open the options list in your conversation with Grace Hopper and Ada Lovelace\n"
-            + self.THREAD
-        )
-        assert strip_conversation_chrome(text) == self.THREAD
-
-    def test_missing_thread_header_strips_only_composer(self):
-        text = self.THREAD + "\nMaximize compose field\nOpen send options"
-        assert strip_conversation_chrome(text) == self.THREAD
-
-    def test_quoted_composer_string_in_message_survives(self):
-        text = (
-            "Open the options list in your conversation with Grace Hopper and Ada Lovelace\n"
-            "Maximize compose field\n"
-            "is the label I keep seeing\n"
-            "Maximize compose field\n"
-            "Open send options"
-        )
-        assert (
-            strip_conversation_chrome(text)
-            == "Maximize compose field\nis the label I keep seeing"
-        )
-
-    def test_quoted_companion_with_suffix_does_not_confirm_composer(self):
-        text = "Hello!\nMaximize compose field\nOpen send options is what I clicked"
-        assert strip_conversation_chrome(text) == text
-
-    def test_quoted_attach_text_does_not_confirm_composer(self):
-        text = (
-            "Hello!\n"
-            "Maximize compose field\n"
-            "Attach an image to your conversation with Grace is the label I clicked"
-        )
-        assert strip_conversation_chrome(text) == text
-
-    def test_distant_companion_text_does_not_confirm_composer(self):
-        filler = "\n".join(f"message {n}" for n in range(10))
-        text = (
-            "Maximize compose field\n"
-            + filler
-            + "\nOpen send options is what I clicked"
-        )
-        assert strip_conversation_chrome(text) == text
-
-    def test_quoted_composer_without_companions_does_not_truncate(self):
-        text = (
-            "Open the options list in your conversation with Grace Hopper and Ada Lovelace\n"
-            "Hello!\n"
-            "Maximize compose field\n"
-            "is what the button says"
-        )
-        assert (
-            strip_conversation_chrome(text)
-            == "Hello!\nMaximize compose field\nis what the button says"
-        )
-
-    def test_quoted_thread_header_in_message_keeps_earlier_messages(self):
-        text = (
-            "Load more conversations\n"
-            "Grace Hopper\n"
-            "Open the options list in your conversation with Grace Hopper and Ada Lovelace\n"
-            "Hello!\n"
-            "Open the options list in your conversation with is a label I quoted\n"
-            "Bye!\n"
-            "Maximize compose field\n"
-            "Open send options"
-        )
-        assert strip_conversation_chrome(text) == (
-            "Hello!\n"
-            "Open the options list in your conversation with is a label I quoted\n"
-            "Bye!"
-        )
-
-    def test_sidebar_end_without_thread_header_still_strips_sidebar(self):
-        text = (
-            "Ada: Preview belonging to a different conversation\n"
-            "Load more conversations\n" + self.THREAD
-        )
-        assert strip_conversation_chrome(text) == self.THREAD
-
-    def test_unknown_locale_returns_unchanged(self):
-        assert strip_conversation_chrome(self.PAGE, locale="de") == self.PAGE
-
-    def test_no_markers_returns_stripped_text(self):
-        assert strip_conversation_chrome("Hello!\nHi there!") == "Hello!\nHi there!"
-
-    def test_empty_string(self):
-        assert strip_conversation_chrome("") == ""
 
 
 class TestActivityFeedExtraction:
@@ -7315,14 +7109,12 @@ class TestMainProfileAlreadyLoaded:
         extractor = LinkedInExtractor(mock_page)
         mock_page.url = "https://www.linkedin.com/in/foo/"
 
-        from linkedin_mcp_server.scraping.extractor import _RATE_LIMITED_MSG
-
         with (
             patch.object(
                 extractor,
                 "_extract_loaded_section",
                 new_callable=AsyncMock,
-                return_value=extracted(_RATE_LIMITED_MSG),
+                return_value=extracted(RATE_LIMITED_SECTION_TEXT),
             ) as loaded,
             patch.object(
                 extractor,
@@ -9834,110 +9626,6 @@ class TestMessageConfirmation:
             _MESSAGE_CONFIRMATION_DISPOSE_JS,
             {"owner": owner, "token": "confirmation-token"},
         )
-
-
-class TestBuildFeedReferences:
-    """Tests for _build_feed_references SDUI-capture / DOM-anchor merging."""
-
-    def test_sdui_urls_become_relative_feed_post_references(self):
-        captured = [
-            "https://www.linkedin.com/posts/alice_some-slug-ugcPost-1-xx",
-            "https://www.linkedin.com/posts/bob_other-post-share-2-yy",
-        ]
-        refs = _build_feed_references([], captured)
-        assert refs == [
-            {
-                "kind": "feed_post",
-                "url": "/posts/alice_some-slug-ugcPost-1-xx",
-                "context": "feed",
-            },
-            {
-                "kind": "feed_post",
-                "url": "/posts/bob_other-post-share-2-yy",
-                "context": "feed",
-            },
-        ]
-
-    def test_duplicate_sdui_urls_are_deduped(self):
-        captured = [
-            "https://www.linkedin.com/posts/alice_x-ugcPost-1-xx",
-            "https://www.linkedin.com/posts/alice_x-ugcPost-1-xx",
-        ]
-        refs = _build_feed_references([], captured)
-        assert len(refs) == 1
-        assert refs[0]["url"] == "/posts/alice_x-ugcPost-1-xx"
-
-    def test_dom_anchor_feed_update_passes_through(self):
-        # DOM anchors that classify_link recognises as feed_post survive
-        # the merge alongside SDUI captures.
-        raw_anchors = [
-            {
-                "href": "https://www.linkedin.com/feed/update/urn:li:activity:1234567890/",
-                "text": "View post",
-            }
-        ]
-        refs = _build_feed_references(raw_anchors, [])
-        assert any(
-            r["url"] == "/feed/update/urn:li:activity:1234567890/"
-            and r["kind"] == "feed_post"
-            for r in refs
-        )
-
-    def test_non_posts_paths_in_sdui_capture_are_skipped(self):
-        # Defensive: only /posts/<slug> shapes count for SDUI append.
-        captured = [
-            "https://www.linkedin.com/in/someuser/",
-            "https://www.linkedin.com/posts/alice_x-ugcPost-1-xx",
-        ]
-        refs = _build_feed_references([], captured)
-        assert [r["url"] for r in refs] == ["/posts/alice_x-ugcPost-1-xx"]
-
-    def test_cap_matches_num_posts_ceiling(self):
-        captured = [
-            f"https://www.linkedin.com/posts/p{i}-ugcPost-{i}-xx" for i in range(60)
-        ]
-        refs = _build_feed_references([], captured)
-        # Cap is 50, mirroring _REFERENCE_CAPS["feed"] / num_posts <= 50.
-        assert len(refs) == 50
-
-    def test_non_feed_post_dom_anchors_are_filtered(self):
-        # Sidebar profile / company / external anchors must not crowd
-        # out SDUI permalinks — references["feed"] is feed_post-only.
-        raw_anchors = [
-            {
-                "href": "https://www.linkedin.com/in/sidebar-user/",
-                "text": "Sidebar User",
-            },
-            {
-                "href": "https://www.linkedin.com/company/some-corp/",
-                "text": "Some Corp",
-            },
-            {
-                "href": "https://example.com/external/",
-                "text": "External Link",
-            },
-        ]
-        refs = _build_feed_references(raw_anchors, [])
-        assert refs == []
-
-    def test_feed_post_dom_anchors_coexist_with_sdui_captures(self):
-        # The two sources fold into the same feed_post kind without
-        # collapsing across URL shapes pointing at the same post.
-        raw_anchors = [
-            {
-                "href": "https://www.linkedin.com/feed/update/urn:li:activity:111/",
-                "text": "View post",
-            }
-        ]
-        captured = ["https://www.linkedin.com/posts/alice_x-ugcPost-1-xx"]
-        refs = _build_feed_references(raw_anchors, captured)
-        urls = [r["url"] for r in refs]
-        kinds = {r["kind"] for r in refs}
-        assert urls == [
-            "/feed/update/urn:li:activity:111/",
-            "/posts/alice_x-ugcPost-1-xx",
-        ]
-        assert kinds == {"feed_post"}
 
 
 class TestDrainListenerTasks:
